requirements.txt
droidrun
llama-index-core
llama-index-llms-google-genai
python-dotenv
google-genai